+++ pkg/cluster/mutate_create_test.go
@@ -32,7 +32,7 @@ func TestClusterCreateMutate(t *testing.T) {
 
 	clusterNetwork := &v1alpha3.ClusterNetwork{
 		APIServerPort: to.Int32Ptr(443),
-		ServiceDomain: "ab123.k8s.test.westeurope.azure.gigantic.io",
+		ServiceDomain: "cluster.local",
 		Services: &v1alpha3.NetworkRanges{
 			CIDRBlocks: []string{
 				"172.31.0.0/16",