+++ pkg/cluster/validate_update_test.go
@@ -23,7 +23,7 @@ func TestClusterUpdateValidate(t *testing.T) {
 
 	clusterNetwork := &v1alpha3.ClusterNetwork{
 		APIServerPort: to.Int32Ptr(443),
-		ServiceDomain: "ab123.k8s.test.westeurope.azure.gigantic.io",
+		ServiceDomain: "cluster.local",
 		Services: &v1alpha3.NetworkRanges{
 			CIDRBlocks: []string{
 				"172.31.0.0/16",
@@ -63,7 +63,7 @@ func TestClusterUpdateValidate(t *testing.T) {
 				"ab123",
 				&v1alpha3.ClusterNetwork{
 					APIServerPort: to.Int32Ptr(80),
-					ServiceDomain: "ab123.k8s.test.westeurope.azure.gigantic.io",
+					ServiceDomain: "cluster.local",
 					Services: &v1alpha3.NetworkRanges{
 						CIDRBlocks: []string{
 							"192.168.0.0/24",
@@ -103,7 +103,7 @@ func TestClusterUpdateValidate(t *testing.T) {
 				"ab123",
 				&v1alpha3.ClusterNetwork{
 					APIServerPort: to.Int32Ptr(443),
-					ServiceDomain: "ab123.k8s.test.westeurope.azure.gigantic.io",
+					ServiceDomain: "cluster.local",
 					Services:      nil,
 				},
 				"api.ab123.k8s.test.westeurope.azure.gigantic.io",