+++ pkg/key/key.go
@@ -11,6 +11,6 @@ func GetControlPlaneEndpointHost(clusterName string, baseDomain string) string {
 	return fmt.Sprintf("api.%s.%s", clusterName, baseDomain)
 }
 
-func ServiceDomain(clusterName string, baseDomain string) string {
-	return fmt.Sprintf("%s.%s", clusterName, baseDomain)
+func ServiceDomain() string {
+	return "cluster.local"
 }